PowerBI Visualization
Visual representation is a core element of the project, and Power BI serves as the ideal tool for this purpose. Power BI's diverse visualization capabilities are leveraged to create interactive and engaging charts and graphs. Each aspect of the employee demographics analysis is paired with suitable visualizations, such as pie charts, bar charts, and maps, to effectively communicate insights to HR stakeholders.